ADC input is now set to 8-bit

+++ Arduino/eTextile_matrix_sensor_blob/collections.h
@@ -96,7 +96,7 @@ void lifo_dequeue(lifo_t *ptr, void* data);
 typedef struct image {
   int w;
   int h;
-  uint16_t* dataPtr;
+  uint8_t* dataPtr;
 } image_t;
 
 ////////////// Bitmap //////////////
